% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from ODE}
\usage{
get_raw_directory()
}
\value{
Raw data frame as downloaded from ODE
}
\description{
Downloads the raw school directory data from the Oregon Department of
Education Report Card API.
}
\keyword{internal}
