% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr_era1}
\alias{process_enr_era1}
\title{Process Era 1 data (2010-2014)}
\usage{
process_enr_era1(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process Era 1 data (2010-2014)
}
\keyword{internal}
