% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_ode_era1}
\alias{read_ode_era1}
\title{Read Era 1 ODE Excel file (2010-2014)}
\usage{
read_ode_era1(file_path, sheet_name, end_year)
}
\arguments{
\item{file_path}{Path to Excel file}

\item{sheet_name}{Name of sheet to read}

\item{end_year}{School year end}
}
\value{
Data frame
}
\description{
Read Era 1 ODE Excel file (2010-2014)
}
\keyword{internal}
