% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_ode_era2}
\alias{read_ode_era2}
\title{Read Era 2 ODE Excel file (2015-present)}
\usage{
read_ode_era2(file_path, sheet_name, end_year)
}
\arguments{
\item{file_path}{Path to Excel file}

\item{sheet_name}{Name of sheet to read}

\item{end_year}{School year end}
}
\value{
Data frame
}
\description{
Read Era 2 ODE Excel file (2015-present)
}
\keyword{internal}
