% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{read_ode_excel}
\alias{read_ode_excel}
\title{Read ODE Excel file with appropriate method for format era}
\usage{
read_ode_excel(file_path, end_year)
}
\arguments{
\item{file_path}{Path to downloaded Excel file}

\item{end_year}{School year end}
}
\value{
Data frame with raw data
}
\description{
Read ODE Excel file with appropriate method for format era
}
\keyword{internal}
