% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_col_names}
\alias{standardize_col_names}
\title{Standardize column names}
\usage{
standardize_col_names(names)
}
\arguments{
\item{names}{Character vector of column names}
}
\value{
Character vector of standardized names
}
\description{
Converts Oregon column names to a standard format for processing.
}
\keyword{internal}
