% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orschooldata-package.R
\docType{package}
\name{orschooldata-package}
\alias{orschooldata-package}
\alias{orschooldata}
\title{orschooldata: Oregon School Data}
\description{
A simple, consistent interface for accessing Oregon school data in Python and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://almartin82.github.io/orschooldata}
  \item \url{https://github.com/almartin82/orschooldata}
  \item Report bugs at \url{https://github.com/almartin82/orschooldata/issues}
}

}
\author{
\strong{Maintainer}: Al Martin \email{almartin@example.com}

}
\keyword{internal}
