% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr_era2}
\alias{process_enr_era2}
\title{Process Era 2 data (2015-present)}
\usage{
process_enr_era2(df, end_year)
}
\arguments{
\item{df}{Raw data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process Era 2 data (2015-present)
}
\keyword{internal}
